+++ config/application.rb
@@ -83,5 +83,8 @@ module Snorby
 
     # Configure sensitive parameters which will be filtered from the log file.
     config.filter_parameters += [:password]
+
+    # SSL
+    # config.force_ssl = true
   end
 end